requirements.txt
pyOpenSSL
dpkt
